+++ examples/standalone-parser/CHANGELOG.md
@@ -1,5 +1,21 @@
 # node-server
 
+## 1.0.43
+
+### Patch Changes
+
+- Updated dependencies
+  - @beff/cli@0.0.46
+  - @beff/client@0.0.46
+
+## 1.0.42
+
+### Patch Changes
+
+- Updated dependencies
+  - @beff/cli@0.0.45
+  - @beff/client@0.0.45
+
 ## 1.0.41
 
 ### Patch Changes