+++ packages/beff-client/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @beff/client
 
+## 0.0.46
+
+### Patch Changes
+
+- better json schema
+- Updated dependencies
+  - @beff/cli@0.0.46
+
+## 0.0.45
+
+### Patch Changes
+
+- Flat json schema export
+- Updated dependencies
+  - @beff/cli@0.0.45
+
 ## 0.0.44
 
 ### Patch Changes